Alterações no Model , Controller , Repository

+++ src/main/java/com/blog/api/controller/PostController.java
@@ -2,6 +2,9 @@ package com.blog.api.controller;
 
 import com.blog.api.model.Post;
 import com.blog.api.repository.PostRepository;
+
+import jakarta.validation.Valid;
+
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
@@ -19,7 +22,7 @@ public class PostController {
     private PostRepository postRepository;
 
     @PostMapping
-    public ResponseEntity<Post> cadastrarPost(@RequestBody Post post) {
+    public ResponseEntity<Post> cadastrarPost(@RequestBody @Valid Post post) {
         return ResponseEntity.status(HttpStatus.CREATED).body(postRepository.save(post));
 
     }
@@ -41,7 +44,7 @@ public class PostController {
     }
 
     @PutMapping("/{id}")
-    public Post atualizarPorId(@PathVariable("id") Long id, @RequestBody Post post) {
+    public Post atualizarPorId(@PathVariable("id") Long id, @RequestBody @Valid Post post) {
         Optional<Post> postExistente = postRepository.findById(id);
 
         if (postExistente.isPresent()) {
@@ -59,10 +62,16 @@ public class PostController {
     }
 
     @DeleteMapping("{id}")
-    public String deletarPost(@PathVariable("id") Long id) {
+    public ResponseEntity<String> deletarPost(@PathVariable("id") Long id) {
+        Optional<Post> post = postRepository.findById(id);
+
+        if (post.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        }
+
         postRepository.deleteById(id);
 
-        return "Post Deletado Com Sucesso!";
+        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
 
     }
